Update FAB design
